requirements.txt
pyTelegramBotAPI==4.0.0
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
